Add files via upload
Dear PW Skills Examiner,
I hope this message finds you well. I am writing to bring to your attention an issue I encountered while attempting to open the SQL format file provided for the assessment.
When I try to open the file directly, I receive errors. However, when I copy the script and run it in SQL Workbench, the script executes without any issues.

I wanted to inform you of this issue to ensure that it does not affect the assessment process. I will make sure to copy the script and run it in SQL Workbench to complete the assessment.